Overwrite our libxls "fork" with current svn "HEAD"
https://github.com/svn2github/libxls/commit/bb84eb995d035dff845f3eec875f0c468f691b26

+++ src/libxls/xlstool.h
@@ -31,7 +31,6 @@
  */
 
 #include "libxls/xlsstruct.h"
-#include "cran.h"
 
 extern void dumpbuf(BYTE* fname,long size,BYTE* buf);
 extern void verbose(char* str);
@@ -48,6 +47,6 @@ extern void xls_showCell(struct st_cell_data* cell);
 extern void xls_showFont(struct st_font_data* font);
 extern void xls_showXF(XF8* xf);
 extern void xls_showFormat(struct st_format_data* format);
-extern BYTE* xls_getfcell(xlsWorkBook* pWB,struct st_cell_data* cell,BYTE *label);
+extern BYTE* xls_getfcell(xlsWorkBook* pWB,struct st_cell_data* cell,DWORD *label);
 extern char* xls_getCSS(xlsWorkBook* pWB);
 extern void xls_showBOF(BOF* bof);
